+++ stories/icon-button.stories.ts
@@ -75,11 +75,10 @@ const metadata: Meta<IgcIconButtonComponent> = {
     },
     disabled: {
       type: 'boolean',
-      description: 'Determines whether the button is disabled.',
+      description: 'The disabled state of the component',
       control: 'boolean',
       defaultValue: false,
     },
-    ariaLabel: { type: 'string', control: 'text' },
     size: {
       type: '"small" | "medium" | "large"',
       description: 'Determines the size of the component.',
@@ -120,9 +119,8 @@ interface IgcIconButtonArgs {
    * See https://developer.mozilla.org/en-US/docs/Web/HTML/Link_types
    */
   rel: string;
-  /** Determines whether the button is disabled. */
+  /** The disabled state of the component */
   disabled: boolean;
-  ariaLabel: string;
   /** Determines the size of the component. */
   size: 'small' | 'medium' | 'large';
 }