+++ CHANGELOG.md
@@ -5,9 +5,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
 ### Added
+- The following components are now Form Associated Custom Elements and support native form element:
+    - Checkbox
+    - Combo
+    - DateTime input
+    - Input
+    - Masked input
+    - Radio
+    - Rating
+    - Single slider
+    - Select
+    - Switch
 - Text area component
+
+### Deprecated
+ - IgcForm component is deprecated
 
 ## [4.3.1] - 2023-08-02
 ### Added